+++ src/formulation/rawmat.cpp
@@ -28,13 +28,13 @@ void rawmat::zeroentries(intdensematrix entriestozero, bool zerorows, bool zeroc
 {
 	int* entriestozeroptr = entriestozero.getvalues();
 
-	long int numentries = entriestozero.count();
+	int numentries = entriestozero.count();
 	if (numentries > 0)
 	{
 		// First build a vector for direct access:
 		std::vector<bool> istozero(mydofmanager->countdofs(), false);
 
-		for (long int i = 0; i < numentries; i++)
+		for (int i = 0; i < numentries; i++)
 			istozero[entriestozeroptr[i]] = true;
 			
 		if (zerorows)
@@ -42,7 +42,7 @@ void rawmat::zeroentries(intdensematrix entriestozero, bool zerorows, bool zeroc
 			for (int i = 0; i < accumulatedrowindices.size(); i++)
 			{
 				int* accumulatedrowindicesptr = accumulatedrowindices[i].getvalues();
-				for (long int j = 0; j < accumulatedrowindices[i].count(); j++)
+				for (int j = 0; j < accumulatedrowindices[i].count(); j++)
 				{
 					if (accumulatedrowindicesptr[j] >= 0 && istozero[accumulatedrowindicesptr[j]])
 						accumulatedrowindicesptr[j] = -1;
@@ -54,7 +54,7 @@ void rawmat::zeroentries(intdensematrix entriestozero, bool zerorows, bool zeroc
 			for (int i = 0; i < accumulatedcolindices.size(); i++)
 			{
 				int* accumulatedcolindicesptr = accumulatedcolindices[i].getvalues();
-				for (long int j = 0; j < accumulatedcolindices[i].count(); j++)
+				for (int j = 0; j < accumulatedcolindices[i].count(); j++)
 				{
 					if (accumulatedcolindicesptr[j] >= 0 && istozero[accumulatedcolindicesptr[j]])
 						accumulatedcolindicesptr[j] = -1;
@@ -113,7 +113,7 @@ void rawmat::removeconstraints(void)
     int* myrows = petscrows.getvalues();
     int* mycols = petsccols.getvalues();
     
-    for (long int i = 0; i < petscrows.count(); i++)
+    for (int i = 0; i < petscrows.count(); i++)
     {
         myrows[i] = dofrenumbering[myrows[i]];
         mycols[i] = dofrenumbering[mycols[i]];
@@ -136,13 +136,13 @@ void rawmat::process(void)
 {
     // Concatenate all accumulated fragments!
     // Remove negative indexes. First get the overall length.
-    long int veclen = 0;
+    int veclen = 0;
     for (int i = 0; i < accumulatedvals.size(); i++)
     {
         int* accumulatedrowindicesptr = accumulatedrowindices[i].getvalues();
         int* accumulatedcolindicesptr = accumulatedcolindices[i].getvalues();
         
-        for (long int j = 0; j < accumulatedvals[i].count(); j++)
+        for (int j = 0; j < accumulatedvals[i].count(); j++)
         {
             if (accumulatedrowindicesptr[j] >= 0 && accumulatedcolindicesptr[j] >= 0)
                 veclen++;            
@@ -154,14 +154,14 @@ void rawmat::process(void)
     int* stitchedcolindices = new int[veclen];
     double* stitchedvals = new double[veclen];
 
-    long int ind = 0;
+    int ind = 0;
     for (int i = 0; i < accumulatedvals.size(); i++)
     {
         double* valsptr = accumulatedvals[i].getvalues();
         int* accumulatedrowindicesptr = accumulatedrowindices[i].getvalues();
         int* accumulatedcolindicesptr = accumulatedcolindices[i].getvalues();
         
-        for (long int j = 0; j < accumulatedvals[i].count(); j++)
+        for (int j = 0; j < accumulatedvals[i].count(); j++)
         {
             if (accumulatedrowindicesptr[j] >= 0 && accumulatedcolindicesptr[j] >= 0)
             {
@@ -175,13 +175,13 @@ void rawmat::process(void)
     
     // Sort the vectors according to the row then according to the column. 
     // 'reorderingvector' will tell us how to reorder.
-    long int* reorderingvector = myalgorithm::stablesortparallel({stitchedrowindices, stitchedcolindices}, veclen);
+    int* reorderingvector = myalgorithm::stablesortparallel({stitchedrowindices, stitchedcolindices}, veclen);
         
     // Get the number of nonzeros:
     nnz = 0;
     if (veclen > 0)
     	nnz = 1;
-    for (long int i = 1; i < veclen; i++)
+    for (int i = 1; i < veclen; i++)
     {
         if (stitchedrowindices[reorderingvector[i]] != stitchedrowindices[reorderingvector[i-1]] || stitchedcolindices[reorderingvector[i]] != stitchedcolindices[reorderingvector[i-1]])
             nnz++;
@@ -198,7 +198,7 @@ void rawmat::process(void)
 
     int row = -1; 
     ind = -1;
-    for (long int i = 0; i < veclen; i++)
+    for (int i = 0; i < veclen; i++)
     {
         // Same row:
         if (i > 0 && stitchedrowindices[reorderingvector[i]] == stitchedrowindices[reorderingvector[i-1]])
